!refactor(*): change folder structure
