+++ client/src/components/Events/Events.js
@@ -5,7 +5,7 @@ import "react-responsive-carousel/lib/styles/carousel.min.css";
 import classes from "./Events.module.css";
 import comingSoonIMG from "../../assets/images/events/coming-soon.png";
 import Rlogo from "../../assets/images/events/Rlogo.png";
-import Carousel from "../Carousel/Carousel";
+import Carousel from "../UI/Carousel/Carousel";
 
 const Events = () => {
   return (
