+++ client/src/components/Index/Focus/Focus.module.css
